requirements.txt
pywebview==5.5.1
pymupdf==1.26.0
rapidfuzz==3.13.0
openpyxl==3.1.5
pyinstaller==6.12.0
